==> preprocessing/01_variance_filtering_tcga_exp.Rmd <==
---
title: "preprocessing_tcga_exp - FINISHED"
author: "Fabian Strobel"
date: '2022-05-06'
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load original data (60,000x10,000 Matrix)
```{r Load data}
## Set directory
setwd('../')
directory = "data/tcga_tumor_log2TPM.RDS"
tcga_exp <- readRDS(directory)
```

## Checking for NAs and removing rows containing NAs
```{r NA check}
# Go through each row and sum up all missing values
#rmv.rows = apply(tcga_exp,1,function(x){sum(is.na(x))})
# Removing any row with 1 or more missing values
#tcga_exp_cleaned = tcga_exp[- which(rmv.rows > 0),]
# Check result
#dim(tcga_exp)
```


```{r means}
## Compute minimal value of matrix
min(tcga_exp)
## The minimal value is -9.9658 -> this equals a read amount of zero

## Compute mean of every row/ for every gene
means = apply(tcga_exp,1,mean)

mean(means)
median(means)
q <- quantile(means, seq(0,1, 0.1))

## Distribution of means of all genes
plot(means, main = "Distribution of all means of tcga_exp"); abline(h = c(q), col="red", lty = 2)
```


```{r spread}
sd <- apply(tcga_exp, 1, sd)
var <- sd^2

hist(sd, main = "Histogramm for SD of each row"); abline(v = median(sd), col = "red")
hist(var, main = "Histogramm for variance of each row"); abline(v = median(var), col = "red")

quant = quantile(var, seq(0, 1, 0.1)) 
hist(var, breaks = 30, main = "Variances for each row of tcga_exp \n and quantiles in 10 % steps"); abline(v = quant, col = "red", lty = 2)

## remove all genes with variance smaller than 1
genes.remove = which(var < quantile(var, 0.35))
tcga_exp_cleaned = tcga_exp[var > quantile(var, 0.35),]

## Amount and share of removed genes
length(genes.remove) ## 21174
share <- length(genes.remove)/nrow(tcga_exp)
quantile(var, share)

## controll
var_cleaned <- apply(tcga_exp_cleaned, 1, var)
min(var_cleaned)
nrow(tcga_exp_cleaned) ## 39324
```

```{r export}
#saveRDS(tcga_exp_cleaned, file = "tcga_exp_variance_filtered")
```
