==> preprocessing/02_hallmarks_genesets_ensembl_IDs.Rmd <==
---
title: "02_hallmarks_genesets_ensembl_IDs"
author: "Fabian Strobel"
date: '2022-05-19'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
setwd('.')
hallmark_genesets_symbol <- readRDS("data/hallmarks_genesets")
```

```{r}
library(EnsDb.Hsapiens.v79)

## Convert from gene.symbol to ensembl.gene
Angio_AACR <- ensembldb::select(EnsDb.Hsapiens.v79, keys= hallmark_genesets_symbol[["genesets"]][["Angio_AACR"]], keytype = "SYMBOL", columns = c("SYMBOL","GENEID"))

genesets <- hallmark_genesets_symbol[["genesets"]]

ens_id <- lapply(genesets, function(genesets){
get_ensd = ensembldb::select(EnsDb.Hsapiens.v79, keys= genesets[], keytype = "SYMBOL", columns = c("SYMBOL","GENEID"))
return(get_ensd)})
```

```{r}
## single example how it should look like
Angio_ALL <- ens_id[["Angio_ALL"]][["GENEID"]]
ens_id_list["Angio_ALL"] <- list(Angio_ALL)

ens_id_list <- lapply(ens_id, function(geneset){
test <- geneset$GENEID
return(test)})
```

```{r}
#saveRDS(ens_id_list, file = "hallmarks_genesets_ensID.rds")
```
