==> preprocessing/04_comparison_of_gene_sets.Rmd <==
---
title: "04_comparison_of_gene_sets"
author: "Fabian Strobel"
date: '2022-05-13'
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load data}
setwd('../')
directory = "data/total_pathways_ensID.rds"
pathways <- readRDS(directory)
```
"The Jaccard Index is a statistical measure that is frequently used to compare the similarity of binary variable sets. It is the length of the union divided by the size of the intersection between the sets."

Jaccard Similarity = (number of observations in both sets) / (number in either set)

$$
J(A, B) = \frac{|A∩B|}{|A∪B|} = \frac{|A∩B|}{|A|+|B|-|A∩B|}
$$
$$
0 \leq J(A, B) \leq 1
$$
Definition of Jaccard Metric
```{r Jaccard function}
jaccard <- function(a, b) {
    intersection = length(intersect(a, b))
    union = length(a) + length(b) - intersection
    return (intersection/union)
}
```


```{r single tests}
jaccard(pathways[["Angio_AACR"]],pathways[["Angio_ALL"]])
jaccard(as.character(pathways[1]), as.character(pathways[2]))
jaccard(pathways[1],pathways[2])

pA <- names(pathways[1])
pB <- names(pathways[2])
jaccard(pathways[[pA]], pathways[[pB]])

class(pathways[["Angio_AACR"]]) # character
class(pathways[1]) # list
```

```{r Jaccard matrix}
jaccard_matrix <- data.frame(matrix(NA, nrow = length(pathways), ncol = length(pathways)))
rownames(jaccard_matrix) = names(pathways)
colnames(jaccard_matrix) = names(pathways)

for (i in 1:length(pathways)) {
  for (j in 1:length(pathways)) {
    x <- jaccard(pathways[[names(pathways[i])]], pathways[[names(pathways[j])]])
    jaccard_matrix[i,j] <- x
  }
}

diag(jaccard_matrix) = NA
```

```{r}
g <- rep("Given Sets", 46)
h <- rep("MSigDB Hallmarks", 50)
k <- rep("KEGG", 186)
o <- rep("Oncogenic Sigantures", 189)
t <- rep("Ontology", 84)

annot <- as.data.frame(c(g, h, k, o, t))

colnames(annot) <- "Gene sets"
rownames(annot) <- names(pathways)
```

```{r heatmap}
library(pheatmap)
library(randomcoloR)

pheatmap(jaccard_matrix, 
         cluster_rows = F, 
         cluster_cols = F,
         main = "Jaccard Indices to compare similarity between different pathways",
         show_rownames = F,
         show_colnames = F,
         annotation_col = annot)
```

```{r}
pheatmap(jaccard_matrix, 
         cluster_rows = T, 
         cluster_cols = T,
         main = "Jaccard Indices to compare similarity between different pathways",
         show_rownames = F,
         show_colnames = F,
         treeheight_row = 0,
         treeheight_col = 0,
         annotation_row = annot)
```
