==> preprocessing/05_biotype_analysis.Rmd <==
---
title: "05_biotype_analysis"
author: "Fabian Strobel"
date: '2022-05-15'
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

-----------------------------------------
## TCGA Matrix ##
-----------------------------------------

```{r genes from expression matrix}
setwd('../')

tcga_exp_cleaned <- readRDS("data/tcga_exp_variance_filtered.rds")
genes <- rownames(tcga_pancancer)
#saveRDS(genes, file = "tcga_exp_variance_filtered_genes.rds")

genes <- readRDS("data/tcga_exp_variance_filtered_genes.rds")
```

```{r Split ENSG identifier from Gene Symbol}
gene_identifier <- strsplit(genes, split = "|", fixed = TRUE)

genes_ens<- sapply(gene_identifier, FUN = function(gene_identifier){
  ens = gene_identifier[1]
  return(ens)
})

genes_symbol<- sapply(gene_identifier, FUN = function(gene_identifier){
  ens = gene_identifier[2]
  return(ens)
})
```

```{r Remove version number}
genes_ens_vn <- strsplit(genes_ens, split = ".", fixed = TRUE)
genes_ensembl <- sapply(genes_ens_vn, FUN = function(genes_ens_vn){
  ens = genes_ens_vn[1]
  return(ens)
})
```

```{r}
rm(gene_identifier)
rm(genes_ens_vn)
rm(genes_ens)
rm(genes_symbol)
```

```{r library}
library(biomaRt)
```

```{r Biotypes of genes}
#mart = useEnsembl(dataset = "hsapiens_gene_ensembl", biomart='ensembl')

#list_biotype = getBM(attributes = c("ensembl_gene_id", "gene_biotype"), filters = "ensembl_gene_id", values = genes_ensembl, mart = mart, useCache = FALSE)
```
38268 matches -> not used

```{r}
library(EnsDb.Hsapiens.v79)
list_biotype <- ensembldb::select(EnsDb.Hsapiens.v79, keys= genes_ensembl, keytype = "GENEID", columns = c("GENEBIOTYPE"))
```
From 39324 input genes, 39226 could be matched with a corresponding biotype. But what about the rest?

```{r}
list_biotype_genes <- list_biotype[,1]

unmatched_genes <- setdiff(genes_ensembl, list_biotype_genes)
```

```{r investigate unmatched data}
unmatched_genes_exp <- tcga_exp_cleaned[unmatched_genes,]

## Compute mean of every row/ for every gene
means_umatched_genes = apply(unmatched_genes_exp,1,mean)
variance_unmatched_genes = apply(unmatched_genes_exp,1,var)

plot(means_umatched_genes,variance_unmatched_genes, main = "Mean-variance plot for genes \n without known biotype")

mv <- data.frame(means_umatched_genes, variance_unmatched_genes)

library(ggplot2)

ggplot(mv, aes(x = means_umatched_genes, y = variance_unmatched_genes)) +
  geom_point() + ggtitle("Mean-variance plot for genes without known biotype")
```

```{r}
biotypes <- unique(list_biotype[2])
biotypes

biotypes_counts <- data.frame(matrix(NA, nrow = nrow(biotypes)))
rownames(biotypes_counts) = biotypes[,1]
colnames(biotypes_counts) = "counts"

for (i in 1:nrow(biotypes)){
  biotypes_counts[i,] <- length(which(list_biotype$GENEBIOTYPE == biotypes[i,]))
}

biotypes_counts

TCGA_biotypes <- biotypes_counts
```



-----------------------------------------
## Gene sets ##
-----------------------------------------
```{r load gene sets}
setwd('../')
total_pathways_ensID <- readRDS("total_pathways_ensID.rds")

p_genes <- unlist(total_pathways_ensID, use.names = F)

length(p_genes)

dupl <- duplicated(p_genes)
table(dupl)
p_genes <- p_genes[dupl == "FALSE"]

length(p_genes)
rm(dupl)
```

```{r library}
library(EnsDb.Hsapiens.v79)
```

```{r Biotypes of genes biomart}
#mart = useEnsembl(dataset = "hsapiens_gene_ensembl", biomart='ensembl')

#p_list_biotype = getBM(attributes = c("ensembl_gene_id", "gene_biotype"), filters = "ensembl_gene_id", values = p_genes, mart = mart, useCache = FALSE)
```
From 8,239 input genes, 7,797 could be matched with a corresponding biotype. But what about the rest?

```{r EnsDb.Hsapiens.v79}
p_list_biotypes_EnsDb <- ensembldb::select(EnsDb.Hsapiens.v79, keys= p_genes, keytype = "GENEID", columns = c("GENEBIOTYPE"))
```
From 23.216 input genes, 22,085 could be matched with a corresponding biotype.

But what about the rest?
```{r}
p_list_biotypes_EnsDb_genes <- p_list_biotypes_EnsDb[,1]

p_unmatched_genes <- setdiff(p_genes, p_list_biotypes_EnsDb_genes)
length(p_unmatched_genes)
```


```{r}
p_biotypes <- unique(p_list_biotypes_EnsDb[2])
p_biotypes

p_biotypes_counts <- data.frame(matrix(NA, nrow = nrow(p_biotypes)))
rownames(p_biotypes_counts) = p_biotypes[,1]
colnames(p_biotypes_counts) = "counts"

for (i in 1:nrow(p_biotypes)){
  p_biotypes_counts[i,] <- length(which(p_list_biotypes_EnsDb$GENEBIOTYPE == p_biotypes[i,]))
}

p_biotypes_counts

pathways_biotypes <- p_biotypes_counts
```


What we can see is that the protein coding genes make up the majority of our pathway genes

Let's extract the protein coding genes of our pathways and store them in *protein_coding*

```{r}
protein_coding = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "protein_coding"), 1]
```

In order to check back on the non coding genes, storing them in an individual variable and combining these variables in *non_coding_pathway_genes* might be helpful

```{r}
miRNA = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "miRNA"), 1]
length(miRNA)

polymorphic_pseudogene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "polymorphic_pseudogene"), 1]
length(polymorphic_pseudogene)

Mt_rRNA = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "Mt_rRNA"), 1]
length(Mt_rRNA)

lncRNA = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "lncRNA"), 1]
length(lncRNA)

processed_pseudogene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "processed_pseudogene"), 1]
length(miRNA)

unprocessed_pseudogene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "unprocessed_pseudogene"), 1]
length(unprocessed_pseudogene)

IG_C_gene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "IG_C_gene"), 1]
length(IG_C_gene)

IG_V_gene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "IG_V_gene"), 1]
length(IG_V_gene)

TR_V_gene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "TR_V_gene"), 1]
length(TR_V_gene)

IG_C_pseudogene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "IG_C_pseudogene"), 1]
length(IG_C_pseudogene)

TR_C_gene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "TR_C_gene"), 1]
length(TR_C_gene)

transcribed_unitary_pseudogene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "transcribed_unitary_pseudogene"), 1]
length(transcribed_unitary_pseudogene)

transcribed_unprocessed_pseudogene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "transcribed_unprocessed_pseudogene"), 1]
length(transcribed_unprocessed_pseudogene)

transcribed_processed_pseudogene = p_list_biotype[which(p_list_biotypes_EnsDb$GENEBIOTYPE == "transcribed_processed_pseudogene"), 1]
length(transcribed_processed_pseudogene)
```

Let's combine and store them in the variable *non_coding_pathway_genes*

```{r}
non_coding_pathway_genes_ens = c(miRNA, polymorphic_pseudogene, Mt_rRNA, lncRNA, processed_pseudogene, unprocessed_pseudogene, IG_C_gene, IG_V_gene, TR_V_gene, IG_C_pseudogene, TR_C_gene, transcribed_unitary_pseudogene, transcribed_unprocessed_pseudogene, transcribed_processed_pseudogene)
```

Before disregarding any of these non coding genes, checking whether any of these genes can be found in our own dataset should be done

```{r}
matches = match(non_coding_pathway_genes_ens, genes_ensembl)
length(matches)


matched_non_coding_pathway_genes = non_coding_pathway_genes_ens[na.omit(matches)]
length(matched_non_coding_pathway_genes) # we end up with 53 matched genes
```

In order to perform a PCA later on, reducing tcga_exp_cleaned by only looking at the matched non coding pathway genes, the protein coding pathway genes, and the genes of our expressed genes that couldn't be defined by a biotype might be helpful

```{r}
tcga_exp_cleaned_reduced = tcga_exp_cleaned[c(unmatched_genes, matched_non_coding_pathway_genes, protein_coding),]
dim(tcga_exp_cleaned_reduced) # we end up with 8766 genes
```

checking for duplicated genes (should be all unique though)

```{r}
duplicates = duplicated(rownames(tcga_exp_cleaned_reduced))
table(duplicates)
```

```{r}
#saveRDS(tcga_exp_cleaned_reduced, file = "tcga_exp_x_total_pathways.rds")
```


***

Explanation of biotypes:    
- Protein coding: Gene/transcipt that contains an open reading frame (ORF)

```{r}
TCGA_biotypes
pathways_biotypes
```

For pan-cancer analysis, we take the most interesting biotypes:
- Short non-coding RNAs, of which some are already well researched and found in bibliographies (Grundvorlesung Biologie 1.FS)
- Long non-coding RNA, which might have regulatory functions
- Protein-coding genes (duh, obviously)
```{r}
## ncRNA: A non-coding gene
TCGA_snRNA <- list_biotype[which(list_biotype$GENEBIOTYPE == "snRNA"), 1]
length(TCGA_snRNA)
TCGA_miRNA <- list_biotype[which(list_biotype$GENEBIOTYPE == "miRNA"), 1]
length(TCGA_miRNA)
TCGA_rRNA <- list_biotype[which(list_biotype$GENEBIOTYPE == "rRNA"), 1]
length(TCGA_rRNA)
TCGA_snoRNA <- list_biotype[which(list_biotype$GENEBIOTYPE == "snoRNA"), 1]
length(TCGA_snoRNA)

##Long non-coding RNA: A non-coding gene/transcript >200bp in length
TCGA_lincRNA <- list_biotype[which(list_biotype$GENEBIOTYPE == "lincRNA"), 1]
length(TCGA_lincRNA)

## Protein coding: Gene/transcipt that contains an open reading frame (ORF) and antibody/Tcell rezeptors
TCGA_protein_coding <-  list_biotype[which(list_biotype$GENEBIOTYPE == "protein_coding"), 1]
length(TCGA_protein_coding)

TCGA_IGVgene <- list_biotype[which(list_biotype$GENEBIOTYPE == "IG_V_gene"), 1]
length(TCGA_IGVgene)

TCGA_IGCgene <- list_biotype[which(list_biotype$GENEBIOTYPE == "IG_C_gene"), 1]
length(TCGA_IGCgene)

TCGA_IGJgene <- list_biotype[which(list_biotype$GENEBIOTYPE == "IG_J_gene"), 1]
length(TCGA_IGJgene)

TCGA_TRVgene <- list_biotype[which(list_biotype$GENEBIOTYPE == "TR_V_gene"), 1]
length(TCGA_TRVgene)

TCGA_TRCgene <- list_biotype[which(list_biotype$GENEBIOTYPE == "TR_C_gene"), 1]
length(TCGA_TRCgene)
```

The genes are now fused into a final dataframe for pan-cancer analysis:
```{r}
TCGA_pancancer_genes <- c(TCGA_snRNA, TCGA_miRNA, TCGA_rRNA, TCGA_snoRNA, TCGA_lincRNA, TCGA_protein_coding, TCGA_IGCgene, TCGA_IGVgene, TCGA_IGJgene, TCGA_TRCgene, TCGA_TRVgene)

length(TCGA_pancancer_genes)
```


```{r}
tcga_pancancer <- tcga_exp_cleaned[TCGA_pancancer_genes,]
table(duplicated(rownames(tcga_pancancer)))
dim(tcga_pancancer)

rownames(tcga_pancancer) <- TCGA_pancancer_genes
```

```{r}
saveRDS(tcga_pancancer, file = "tcga_pancancer.rds")
```

```{r}
# Other genes (not exported)
#all_genes <- rownames(tcga_exp_variance_filtered)
#other_genes <- setdiff(all_genes, TCGA_pancancer_genes)

#tcga_pancancer_excluded <- tcga_exp_variance_filtered[other_genes,]

#saveRDS(tcga_pancancer_excluded, file = "tcga_pancancer_excluded")
```
