==> preprocessing/06_venn_diagramm.Rmd <==
---
title: "Untitled"
author: "Fabian Strobel"
date: '2022-06-22'
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r}
setwd('..')
pathways <- readRDS("data/total_pathways_ensID.rds")
tcga_pancancer <- readRDS("data/tcga_pancancer.rds")
focussed_analysis <- readRDS("data/PRAD_TvsN.rds")
library(ggvenn)
library(ggVennDiagram)
library(ggplot2)
```

Pathway genes
```{r}
## Combine all pathway genes into one character
pathway_genes <- unlist(pathways, use.names = F)

## Remove duplicates
pathway_genes_uni <- unique(pathway_genes)
rm(pathway_genes)
```

***
### PANCANCER

TCGA genes
```{r}
tcga_genes <- rownames(tcga_pancancer)
```

Compute the intersection
```{r}
length(intersect(tcga_genes, pathway_genes_uni))
```
Venn Diagram
```{r}
x <- list(RNAseq = tcga_genes, Pathways = pathway_genes_uni)

ggvenn(x) + 
  ggtitle("Venn diagram of TCGA RNAseq genes and pathway genes") +
  theme(plot.title = element_text(hjust = 0.5))

ggVennDiagram(x) +
  ggtitle("Venn diagram of TCGA RNAseq genes and pathway genes") +
  theme(plot.title = element_text(hjust = 0.5)) +
  ggplot2::scale_fill_gradient(low="lightsalmon",high = "lightblue2")
```

Determining the genes with the highes variance and their overlap with the pathways
```{r}
# Computing the variance
v <- apply(tcga_pancancer, 1, var) 

# Getting genes with highest variance
topVar <- tcga_pancancer[v > quantile(v, 0.90),]
```

```{r}
topVar_genes <- rownames(topVar)

x <- list(RNAseqTopVar = topVar_genes, Pathways = pathway_genes_uni)

ggVennDiagram(x) +
  ggtitle("TCGA RNAseq genes with var > q90") +
  theme(plot.title = element_text(hjust = 0.5)) +
  ggplot2::scale_fill_gradient(low="lightsalmon",high = "lightblue2")
```


***
### FOCUSED ANALYSIS

TCGA genes
```{r}
prad_genes <- rownames(focussed_analysis$tumor)
```

Compute the intersection
```{r}
length(intersect(prad_genes, pathway_genes_uni))
```
Venn Diagram
```{r}
x <- list(RNAseq = prad_genes, Pathways = pathway_genes_uni)

ggvenn(x) + 
  ggtitle("Venn diagram of PRAD RNAseq genes and pathway genes") +
  theme(plot.title = element_text(hjust = 0.5))

ggVennDiagram(x) +
  ggtitle("Venn diagram of PRAD RNAseq genes and pathway genes") +
  theme(plot.title = element_text(hjust = 0.5)) +
  ggplot2::scale_fill_gradient(low="lightsalmon",high = "lightblue2")
```
